+++ Locadora_V2/src/Forms/JFrmDiretor.java
@@ -4,6 +4,8 @@ import ControlersDao.DiretorDao;
 import Model.DiretorBean;
 import java.awt.HeadlessException;
 import java.util.ArrayList;
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
 
@@ -17,24 +19,27 @@ public class JFrmDiretor extends javax.swing.JFrame {
     
     DefaultTableModel model;
     int selected_index = -1;
-    boolean is_update = false;
     
     public JFrmDiretor() {
         initComponents();
         setLocationRelativeTo(null);
         dao = new DiretorDao();
+        diretor = new DiretorBean();
     }
 
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
+        jpmMenu = new javax.swing.JPopupMenu();
+        jmAlterar = new javax.swing.JMenuItem();
         jPanel3 = new javax.swing.JPanel();
         jLabel3 = new javax.swing.JLabel();
         jPanel2 = new javax.swing.JPanel();
         jbtnDelete = new javax.swing.JButton();
         jbtnSave = new javax.swing.JButton();
         jbtnExit = new javax.swing.JButton();
+        jbtLimpar = new javax.swing.JButton();
         jScrollPane1 = new javax.swing.JScrollPane();
         jtbLista = new javax.swing.JTable();
         jPanel1 = new javax.swing.JPanel();
@@ -43,6 +48,15 @@ public class JFrmDiretor extends javax.swing.JFrame {
         jtxtId = new javax.swing.JTextField();
         jtxtNome = new javax.swing.JTextField();
         jcbSituacao = new javax.swing.JCheckBox();
+
+        jmAlterar.setIcon(new javax.swing.ImageIcon(getClass().getResource("/Media/Icons/available_updates-26.png"))); // NOI18N
+        jmAlterar.setText("Alterar");
+        jmAlterar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jmAlterarActionPerformed(evt);
+            }
+        });
+        jpmMenu.add(jmAlterar);
 
         setDefaultCloseOperation(javax.swing.WindowConstants.DO_NOTHING_ON_CLOSE);
         addWindowListener(new java.awt.event.WindowAdapter() {
@@ -100,6 +114,13 @@ public class JFrmDiretor extends javax.swing.JFrame {
             }
         });
 
+        jbtLimpar.setText("Limpar");
+        jbtLimpar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jbtLimparActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
         jPanel2Layout.setHorizontalGroup(
@@ -108,6 +129,8 @@ public class JFrmDiretor extends javax.swing.JFrame {
                 .addContainerGap()
                 .addComponent(jbtnExit)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jbtLimpar)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jbtnSave)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                 .addComponent(jbtnDelete)
@@ -117,12 +140,16 @@ public class JFrmDiretor extends javax.swing.JFrame {
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                    .addComponent(jbtnDelete, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(jbtnSave, javax.swing.GroupLayout.DEFAULT_SIZE, 41, Short.MAX_VALUE)
-                        .addComponent(jbtnExit, javax.swing.GroupLayout.DEFAULT_SIZE, 41, Short.MAX_VALUE)))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPanel2Layout.createSequentialGroup()
+                        .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                            .addComponent(jbtnDelete, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                            .addGroup(javax.swing.GroupLayout.Alignment.LEADING, jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                .addComponent(jbtnSave, javax.swing.GroupLayout.DEFAULT_SIZE, 41, Short.MAX_VALUE)
+                                .addComponent(jbtnExit, javax.swing.GroupLayout.DEFAULT_SIZE, 41, Short.MAX_VALUE)))
+                        .addGap(0, 0, Short.MAX_VALUE))
+                    .addComponent(jbtLimpar, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
         );
 
         jtbLista.setModel(new javax.swing.table.DefaultTableModel(
@@ -142,11 +169,7 @@ public class JFrmDiretor extends javax.swing.JFrame {
                 "Código", "Nome"
             }
         ));
-        jtbLista.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jtbListaMouseClicked(evt);
-            }
-        });
+        jtbLista.setComponentPopupMenu(jpmMenu);
         jScrollPane1.setViewportView(jtbLista);
 
         jPanel1.setBorder(javax.swing.BorderFactory.createEtchedBorder());
@@ -221,50 +244,36 @@ public class JFrmDiretor extends javax.swing.JFrame {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    private void jtbListaMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jtbListaMouseClicked
-        selected_index = jtbLista.getSelectedRow();
-
-        if(selected_index != -1) {
-            diretor = list.get(selected_index);
-            jtxtId.setText(String.valueOf(diretor.getId()));
-            jtxtNome.setText(diretor.getNome());
-            jcbSituacao.setSelected(diretor.getSitucacao() == 1);
-        }
-    }//GEN-LAST:event_jtbListaMouseClicked
-
     private void jbtnDeleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtnDeleteActionPerformed
-        int id = Integer.valueOf(jtxtId.getText());
         selected_index = jtbLista.getSelectedRow();
-        
-        if (selected_index != -1) {
-            if (JOptionPane.showConfirmDialog(this, "Deseja realmente excluir o autor?", null, JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION) {
-                list.remove(selected_index);
-
-                try {
+        if(!jtxtId.getText().isEmpty()) {
+            if (selected_index != -1) {
+                if (JOptionPane.showConfirmDialog(this, "Deseja realmente excluir o autor?", null, JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION) {
+                    list.remove(selected_index);
+                    
                     diretor = new DiretorBean();
-                    diretor.setId(id);
+                    diretor.setId(Integer.valueOf(jtxtId.getText()));
                     dao.delete(diretor);
                     loadTable();
                     clear();
-                    
                     JOptionPane.showMessageDialog(null, "Registro excluido com sucesso!");
-                    
-                } catch (HeadlessException ex) {
+                }
+                else {
                     JOptionPane.showMessageDialog(null, "Erro ao excluir este registro!");
                 }
             }
+        } else {
+            JOptionPane.showMessageDialog(null, "Selecione um cliente");
         }
     }//GEN-LAST:event_jbtnDeleteActionPerformed
 
     private void jbtnSaveActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtnSaveActionPerformed
-        String msg = ("Salvo com sucesso");
-        
         diretor.setNome(jtxtNome.getText());
         diretor.setSituacao(jcbSituacao.isSelected() == true ? 1 : 2);
         dao.save(diretor);
-        clear();
+        perfumaria();
         loadTable();
-        JOptionPane.showMessageDialog(this, msg);
+        clear();
     }//GEN-LAST:event_jbtnSaveActionPerformed
 
     private void jbtnExitActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtnExitActionPerformed
@@ -272,6 +281,8 @@ public class JFrmDiretor extends javax.swing.JFrame {
     }//GEN-LAST:event_jbtnExitActionPerformed
 
     private void formWindowOpened(java.awt.event.WindowEvent evt) {//GEN-FIRST:event_formWindowOpened
+        jcbSituacao.setVisible(false);
+        jtxtId.setEnabled(false);
         loadTable();
     }//GEN-LAST:event_formWindowOpened
 
@@ -280,6 +291,22 @@ public class JFrmDiretor extends javax.swing.JFrame {
             dispose();
         }
     }//GEN-LAST:event_formWindowClosing
+
+    private void jmAlterarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jmAlterarActionPerformed
+        selected_index = jtbLista.getSelectedRow();
+
+        if(selected_index != -1) {
+            diretor = list.get(selected_index);
+            jtxtId.setText(String.valueOf(diretor.getId()));
+            jtxtNome.setText(diretor.getNome());
+            jcbSituacao.setVisible(true);
+            jcbSituacao.setSelected(diretor.getSitucacao() == 1);
+        }
+    }//GEN-LAST:event_jmAlterarActionPerformed
+
+    private void jbtLimparActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jbtLimparActionPerformed
+        clear();
+    }//GEN-LAST:event_jbtLimparActionPerformed
 //<editor-fold defaultstate="collapsed" desc="Componentes">
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
@@ -290,34 +317,51 @@ public class JFrmDiretor extends javax.swing.JFrame {
     private javax.swing.JPanel jPanel2;
     private javax.swing.JPanel jPanel3;
     private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JButton jbtLimpar;
     private javax.swing.JButton jbtnDelete;
     private javax.swing.JButton jbtnExit;
     private javax.swing.JButton jbtnSave;
     private javax.swing.JCheckBox jcbSituacao;
+    private javax.swing.JMenuItem jmAlterar;
+    private javax.swing.JPopupMenu jpmMenu;
     private javax.swing.JTable jtbLista;
     private javax.swing.JTextField jtxtId;
     private javax.swing.JTextField jtxtNome;
     // End of variables declaration//GEN-END:variables
     //</editor-fold>
+
+    public void perfumaria() {
+        
+        
+        String msg = (selected_index != -1 ? "Alterado com sucesso!" : "Salvo com sucesso");
+        String icon = (selected_index != -1 ? 
+                "E:\\FkBr\\Documentos\\GitHub\\LocadoraJava_V2\\Locadora_V2\\src\\Media\\Icons\\edit-26.png" : 
+                "E:\\FkBr\\Documentos\\GitHub\\LocadoraJava_V2\\Locadora_V2\\src\\Media\\Icons\\checkmark-26.png");
+        ImageIcon Icon = new ImageIcon(icon);
+        JOptionPane.showMessageDialog(this, msg, null, 0, Icon);
+    } 
     
     private void clear(){
         jtxtId.setText(null);
         jtxtNome.setText(null);
-        is_update = false;
+        jcbSituacao.setVisible(false);
         selected_index = -1;
         diretor = new DiretorBean();
     }
 
-     private void loadTable(){
-        Object titulo[] = {"Código", "Nome"};
-        Object grade[][] = null;
-        model = new DefaultTableModel(grade, titulo);
-        jtbLista.setModel(model);
-        this.list = dao.all();
+    private void loadTable(){
+       Object titulo[] = {"Código", "Nome"};
+       Object grade[][] = null;
+       model = new DefaultTableModel(grade, titulo);
+       this.list = dao.all();
+       jtbLista.setModel(model);
 
-        for(DiretorBean diretor: list){
-            Object campos[] = {String.format("%05d", diretor.getId()), diretor.getNome()};
-            model.addRow(campos);
-        }
-    }
+       for(DiretorBean _diretor: list){
+           Object campos[] = {String.format("%05d", _diretor.getId()), _diretor.getNome()};
+           model.addRow(campos);
+       }
+   }
+  
 }
+
+
